Type:fix
Description:1.display the max of speed is 300MB/S 2.display speed and the rest of time when click rightmenu on desktop

RootCause:Null

Solution:1.add speed judgement condition 2.add calculation after click rightmenu on desktop

+++ deepin-compressor/source/inc/progress.h
@@ -53,7 +53,7 @@ public:
 
     //add
     void setspeed(double speed);
-    void setresttime(int resttime);
+    void setresttime(double resttime);
 
 signals:
     void  sigCancelPressed(int compressType);
@@ -79,7 +79,7 @@ private:
     QLabel *m_resttimelabel;
 
     double m_speed;
-    int m_resttime;
+    double m_resttime;
 
     bool isStart = false;
